requirements.txt
pip>=21.2.3
setuptools>=57.4.0
# ------ Adv. Math
numpy
# ------ Image Processing
pillow
opencv-python
# ------ OpenGL
pygame>=2.1.3.dev8
PyOpenGL
PyOpenGL_accelerate
# ------ Vulkan
glfw
vulkan